01_user_service/pyproject.toml
[tool.poetry]
name = "online-store"
version = "0.1.0"
description = ""
authors = ["Ch-Saqib17 <choudary1598@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.111.0"
sqlmodel = "^0.0.18"
uvicorn = "^0.29.0"
pytest = "^8.2.1"
aiokafka = "^0.10.0"
protobuf = "^5.27.0"
passlib = "^1.7.4"
python-dotenv = "^1.0.1"
psycopg2 = "^2.9.9"
bcrypt = "^4.1.3"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
pytest-asyncio = "^0.23.7"
azure-eventhub = "^5.12.1"
azure-eventhub-checkpointstoreblob-aio = "^1.1.4"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
